ARCO-160: Centralize message queue client
