+++ internal/async/mq_client_request_txs.go
@@ -1,0 +1,40 @@
+package async
+
+import (
+	"errors"
+	"fmt"
+
+	"github.com/nats-io/nats.go"
+)
+
+const (
+	RequestTxTopic = "request-tx"
+	requestTxGroup = "request-tx-group"
+)
+
+func (c MQClient) SubscribeRequestTxs() error {
+	if c.requestTxChannel == nil {
+		return errors.New("request txs channel is nil")
+	}
+
+	subscription, err := c.nc.QueueSubscribe(RequestTxTopic, requestTxGroup, func(msg *nats.Msg) {
+		c.requestTxChannel <- msg.Data
+	})
+
+	c.requestSubscription = subscription
+
+	if err != nil {
+		return fmt.Errorf("failed to subscribe to %s topic: %w", RequestTxTopic, err)
+	}
+
+	return nil
+}
+
+func (c MQClient) PublishRequestTx(hash []byte) error {
+	err := c.nc.Publish(RequestTxTopic, hash)
+	if err != nil {
+		return fmt.Errorf("failed to publish on %s topic %w", RequestTxTopic, err)
+	}
+
+	return nil
+}
